Implemented submitting to OBS without creating local copy.
This change implements major redesign of the way gbs works
with OBS. Currently it creates working copy, which requires
a lot of code to support it. This change proposes much more
simple way - without creating local copy. It drastically
simplifies the code (+324, -1470) and illiminates buildservice
and obspkg modules. Fixes #137 and #138

Change-Id: I2fe944246da89b2919b6148abe9713d69558f1e1