cmd_build: Disabled auto patch generation. Fixes #DEVT-9
By default git-buildpackage tries to find a way to generate proper
packaging info from git. For example if it founds branch with the name
'upsream' it's looking for the upstream/version tag, where version is
a version of the upstream sources. If it doesn't found the tag it
fails to export. This behaviour caused confusiong among Tizen
developers, as they don't use proper git layout, but some projects use
upstream branches. This patch should fix this kind of issues. With
--git-no-auto-patch-gen git-buildpackage should ignore all upstream
and vendor tags and create one tarball, without generating patches.
This change also includes pointing to the same commit as an upstream
tree. Otherwise git-buildpackage will be looking at upstream/version
tag if upstream branch exists.

Change-Id: Iba62396fef0f82b46589f8094137c41c7d94895b