Make dependencies to osc and build strict. Make dependency to git-buildpackage-rpm versioning.
As gbs depends on locally modified versions of osc and build packages
it makes sence to strictly depend on them to prevent installation of
the bigger version of the package from distro repos.

In case of git-buildpackage-rpm it's hardly possible for any distro to
include bigger version as we're using upstream package and developing
on top of it, so versioning dependency is enough.

Change-Id: I757708a35362f8a68f81fad472ac540c3286fa92