Implemented cleaning of temporary files and directories.
Implemented in utils.Temp class, which creates temporary file or
directory using tempfile.md[ds]temp and destroys it when object is
deleted(garbage collected when it's out of the scope or upon program
exit).  Fixes #184 and #175.

To avoid abnormal gbs exit(and not deleting temporaries) Exception is
catched in gbs main module.

Change-Id: I04d7067fc293e8b712f183a3d8260dc946058fa7