Use logging from gbp and delete the msger module
As gbs already depends on gbp it makes sense to use its logging facility
instead of msger, which is copy&pasted from another project. Gbp logging
is based on standard Python logging functionality. It basically only
adds color output to it. This change adds a minimal functionality to the
'gitbuildsys.log' module to act as a GBS-specific logger.

This change is also takes more than 400 lines of code from the project
and decreases amount of pylint errors on more than 30% (120->84)

Change-Id: I4bd4d2b9299bf7aa79d5ea9da182312d53e8c3ff
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>