Configure coverage to cover only gbs modules
Previous attempt to configure coverage with setup.cfg was not working
properly. It can be seen in OTC Tools Tester Jenkins job as coverage
report includes gbp and even nose modules. This change is the second
attempt to fix the issue. Expected result is to see only gbs modules in
Jenkins coverage report.

Change-Id: I19c5d4624ce4338a5c6a3d7cb1815bbea5bd05b7
Signed-off-by: Ed Bartosh <eduard.bartosh@intel.com>