export: fix exporting multiple spec files
We need to use the 'force-create' option of gbp. Otherwise gbp will
return an error if pristine-tar checkout fails (pristine-tar branch is
found but the tarball is not committed there), as gbs uses the
'pristine-tar' option.

For subsequent spec files, use the 'no-create-orig' option of gbp to
prevent re-creation of tarballs.

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>