Attempt to bypass HTTP caches when fetching repomd.xml
If a repository and its repomd.xml changes frequently it may look
inconsistent to clients. Unlike other files in a repository repomd.xml
does not change its name along with contents like RPM packages and
metadata files with hashes in filenames. To get the most up-to-date
version additional HTTP headers (Pragma, Cache-Control) need to be passed.

Fixes: DEVT-110

Change-Id: I6a20f829d3a5331f9e99820fa46cdb497a5348ee
Signed-off-by: Łukasz Stelmach <l.stelmach@samsung.com>
Cc: Zhang, Qiang Z <qiang.z.zhang@intel.com>