Enable gbs man page
In current version, `man gbs` is disabled by annotating code that creates
manual page in spec file because the creation of manual page will fail
without question due to syntax error in GBS.rst.

In this change, the syntax error in GBS.rst is fixed, and the annotation
is cancelled accordingly.

Change-Id: Id69b8f704a1aa041a83f908ed65798157ca9c5fe
Signed-off-by: Zhang Qiang <qiang.z.zhang@intel.com>