export: support for the new development/* branches
Export now detects the package maintenance model (separate packaging and
devel branches or the old single packaging branch) from the local
.gbs.conf and available branches. If an orphan branch (no common history
with the upstream branch) patch export is disabled and packaging files
are taken as is from the branch. If gbs detects from the local .gbs.conf
that the current branch is a development branch the packaging files are
taken from the packaging branch, but, patches are generated from the
development branch.

Change-Id: Ic7b782480a36792ae5ae55f8159a6850d894799a
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>