import: add --orphan-packaging option
This option causes packaging files to be imported into an orphan branch
which has no common history with the upstream branch. Patches are not
applied either, but, packaging files are imported "as is". The new
options is for supporting the new package maintenance model which has
separate branches for packaging files and source code development. After
importing a source package with this option, a development branch can be
created with "gbs devel start".

NOTE! The new option is only valid for implementing source rpms - no
effect when importing upstream tarballs.

Change-Id: Ib7004ccacf8bc15a9935ade60910a74c577893cb
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>