changelog: utilize rpm-ch from git-buildpackage
Start to use the gbp changelog tool as backend as it is more
feature-rich. For example, it automatically parses bug-tracking-system
meta-tags from the commit messages (i.e. "Closes: XYZ" etc) and has a
lot more intelligent start-commit guessing.

Later on, additional features from gbp can be enabled in gbs, too.

Creating a new changelog from scratch now requires giving the --since
option. Previously gbs would take the whole git history to the changelog
which is problematic - e.g. creating a new changelog for linux kernel
would take every commit from the linux git history which is *very* long.

Also, the order of entries (i.e. commits) in the new changelog section
is now reversed: the oldest commit will be first (topmost) in the list
and the newest commit will be at the bottom of the list. This should
make the changelog a better chronological read.

Change-Id: Ib4f02f9be95f869b2161d16f8b68f828c5c1deda
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>